requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pycparser==2.21
python-dotenv==1.0.0
python-telegram-bot==12.0.0
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
tornado==6.3
urllib3==1.26.15
yarl==1.8.2
